Synchronize with Feral CVS repository:
I don't know why I thought this would work- forcing a re-login to the fabric
controller just barfs.

Make down count reflect jiffies elapsed rather than depend on calls to
ISP_FC_RESCAN.

Shorten deadloop time default to 10 seconds.

At least enable MSI (if not MSI-X) for 24XX cards.


git-svn-id: http://svn.code.sf.net/p/scst/svn/trunk@282 d57e44dd-8a1f-0410-8b47-8ef2f437770f

+++ qla_isp/linux/isp_linux.h
@@ -1,4 +1,4 @@
-/* $Id: isp_linux.h,v 1.152 2008/01/16 20:33:48 mjacob Exp $ */
+/* $Id: isp_linux.h,v 1.153 2008/01/25 22:22:47 mjacob Exp $ */
 /*
  *  Copyright (c) 1997-2007 by Matthew Jacob
  *  All rights reserved.
@@ -638,7 +638,7 @@ typedef struct {
         blocked                 :   1,
         deadloop                :   1,
         role                    :   2;
-    unsigned int downcount;
+    unsigned long downcount;
     unsigned int qfdelay;
 } isp_data;
